requirements.txt
python-swiftclient==3.6.0
python-keystoneclient==3.18.0
requests
pytz
yaml
